+++ test/embulk/input/test_marketo.rb
@@ -108,7 +108,7 @@ module Embulk
 
         def expected_guessed_columns
           [
-            {name: "Id", type: "integer"},
+            {name: "Id", type: "long"},
             {name: "Email", type: "string"},
             {name: "FieldName", type: "string"},
           ]
